LU Factorization

Doolittle algorithm

We are solving a linear equation

\[\mathrm{ A } \mathbf{ x } = \mathbf{ b },\]

where \mathrm{ A } is an n \times n matrix. Turn it into an equivalent form

\[\mathrm{ U } \mathbf{ x } = \mathbf{ c },\]

where \mathrm{ U } is an upper triangular matrix.

For the kth step, we have an elimination matrix \mathrm{ E }_k, so

\[\mathrm{ U } = \mathrm{ E }_{n - 1} \cdots \mathrm{ E_2 E_1 } \mathrm{ A } \equiv \mathrm{ E A },\]

where \mathrm{ E } = \mathrm{ E }_{n - 1} \cdots \mathrm{ E_2 E_1 }. Thus we have n - 1 steps in total. We denote here

\[\mathrm{ A }_{k} = \mathrm{ E }_k \mathrm{ A }_{k - 1}.\]

Thus the final step will be \mathrm{ A }_{n - 1} = \mathrm{ E }_{n - 1} \mathrm{ A }_{n - 2} = \mathrm{ U }.

So

\[\mathrm{ A } = \mathrm{ E }^{-1} \mathrm{ U } = (\mathrm{ E }_{n - 1} \cdots \mathrm{ E_2 E_1 })^{-1} \mathrm{ U } = \mathrm{ E }^{-1}_1 \mathrm{ E }^{-1}_2 \cdots \mathrm{ E }^{-1}_{n - 1} \mathrm{ U } = \mathrm{ L }_1 \mathrm{ L }_2 \cdots \mathrm{ L }_{n - 1} \mathrm{ U } \equiv \mathrm{ L } \mathrm{ U },\]

where \mathrm{ L } = \mathrm{ L }_1 \mathrm{ L }_2 \cdots \mathrm{ L }_{n - 1}.

APMAE4001.LUFactorization.Doolittle.elimination_factors — Method.

elimination_factors(n::Int, k::Int, A::AbstractMatrix)

Generate a vector of factors to eliminate the kth column. Combine this function with row_vector to to generate an general elementary elimination matrix.

Arguments

n::Integer: the size of the matrix A.

k::Integer: the step count in the elimination process. It must satisfies 1 ≤ k ≤ n - 1.

A::AbstractMatrix: the matrix from (k - 1)th step of the elimination process.

source

APMAE4001.LUFactorization.Doolittle.general_elementary_elimination_matrix — Method.

general_elementary_elimination_matrix(lk::AbstractVector, ek::AbstractVector)

Elimination matrix $E_{k}$ do row operations to eliminate the kth column.

source

APMAE4001.LUFactorization.Doolittle.doolittle_lu — Method.

doolittle_lu(A::AbstractMatrix; include_l::Bool=true, include_e::Bool=false)

Referenced from Wikipedia "LU decomposition" page.

source